--- core/libp2p/crypto/error/CMakeLists.txt
@@ -1,13 +1,0 @@
-#
-# Copyright Soramitsu Co., Ltd. All Rights Reserved.
-# SPDX-License-Identifier: Apache-2.0
-#
-
-add_library(crypto_error
-        error.hpp
-        error.cpp
-        )
-
-target_link_libraries(crypto_error
-        outcome
-        )
