+++ core/libp2p/crypto/impl/crypto_provider_impl.cpp
@@ -5,7 +5,7 @@
 
 #include "libp2p/crypto/impl/crypto_provider_impl.hpp"
 
-#include "libp2p/crypto/error/error.hpp"
+#include "libp2p/crypto/error.hpp"
 #include "libp2p/crypto/private_key.hpp"
 #include "libp2p/crypto/proto/keys.pb.h"
